requirements.txt
pyyaml==6.0.2
numpy==2.1.2
scipy==1.14.1
pandas==2.2.3
matplotlib==3.9.2
mip==1.16rc0
